Refactor: 아토믹디자인에 맞게 폴더구조 리팩토링
