--- todo/src/components/molecules/TodoListWrap/index.jsx
@@ -1,14 +1,0 @@
-import React from "react";
-import styled from "styled-components";
-
-const Wrap = styled.div`
-  display: flex;
-  justify-content: space-between;
-  width: 1110px;
-`;
-
-const TodoListWrap = ({ children, ...props }) => {
-  return <Wrap {...props}>{children}</Wrap>;
-};
-
-export default TodoListWrap;
